requirements.txt
torch
torchvision
fastapi
uvicorn
pillow
numpy
pandas
scikit-learn
tqdm
python-multipart
axios
requests
